Option of segmented active layers
The active layer can be composed of few (up to 5) concetri layers of with different resolution.
E.g. current layer
               name material Zc  DZ     resX   resY  eff  RMin   RMax
activelayer:vt VT3 SILICON   20 400e-4 10.e-4 10.e-4 0.9 0.4444 9.3928
can be redefined to
               name material Zc  DZ     resX   resY  eff  RMin  RMax  resx1 resx2 RMax1 resx2 resx2 RMax2
activelayer:vt VT3 SILICON   20 400e-4 10.e-4 10.e-4 0.9 0.4444 5.00  20e-4 20e-4 7     30e-4 30e-4 9.3928
i.e. to have 10 micron resolution at 0.444<r<5, 20 micron at 5<r<7 and 30 micron at r>7
