+++ files/KMCLayerFwd.cxx
@@ -9,15 +9,12 @@ Double_t KMCLayerFwd::fgDefEff = 1.0;
 KMCLayerFwd::KMCLayerFwd(const char *name) 
   : TNamed(name,name)
   ,fZ(0)
-  ,fRMin(0)
-  ,fRMax(1e9)
   ,fThickness(0)
-  ,fx2X0(0)
-  ,fXRho(0)
-  ,fXRes(0)
-  ,fYRes(0)
-  ,fEff(0)
+  ,fx2X0(0.)
+  ,fXRho(0.)
+  ,fEff(0.)
   ,fIsDead(kFALSE)
+  ,fNAccReg(1)
   ,fType(-1)
   ,fActiveID(-1)
   ,fSig2EstX(999)
@@ -28,6 +25,13 @@ KMCLayerFwd::KMCLayerFwd(const char *name)
   ,fTrCorr()
   ,fTrMC("KMCProbeFwd",5)
 {
+  for (int i=0;i<kMaxAccReg;i++) {
+    fRMin[i] = fRMax[i] = -1;
+    fXRes[i] = fYRes[i] = 0;
+  }
+  fRMin[0] = 0;
+  fRMax[0] = 999.;
+
   Reset();
 }
 
@@ -55,8 +59,12 @@ KMCProbeFwd* KMCLayerFwd::AddMCTrack(KMCProbeFwd* src)
 //__________________________________________________________________________
 void KMCLayerFwd::Print(Option_t *opt) const
 {
-  printf("Lr%3d(A%3d) %15s Z=%+7.1f DZ=%7.3f X2X0=%.3e XRho=%.3e SigX=%.3e SigY=%.3e RMin:%.3e RMax:%.3e Eff:%4.2f\n",
-	 GetUniqueID(),fActiveID,GetName(), fZ, fThickness, fx2X0,fXRho,fXRes,fYRes,fRMin,fRMax,fEff);
+  printf("Lr%3d(A%3d) %15s Z=%+7.1f DZ=%7.3f X2X0=%.3e XRho=%.3e SigX=%.3e SigY=%.3e Eff:%4.2f RMin:%.3e RMax:%.3e ",
+	 GetUniqueID(),fActiveID,GetName(), fZ, fThickness, fx2X0,fXRho,fXRes[0],fYRes[0],fEff,fRMin[0],fRMax[0]);
+  for (int ir=1;ir<fNAccReg;ir++) { // print extra regions
+    printf("SigX=%.3e SigY=%.3e RMax:%.3e ",fXRes[ir],fYRes[ir],fRMax[ir]);
+  }
+  printf("\n");
   TString opts = opt; opts.ToLower();
   //
   if (opts.Contains("cl")) {
@@ -76,4 +84,5 @@ KMCProbeFwd* KMCLayerFwd::GetWinnerMCTrack()
   if (!win || win->IsKilled()) return 0;
   return win;
 }
+
